+++ management/server/idp/auth0idp_test.go
@@ -1,4 +1,4 @@
-package idpmanager
+package idp
 
 import (
 	"encoding/json"
