+++ management/server/idp/util.go
@@ -1,4 +1,4 @@
-package idpmanager
+package idp
 
 import "encoding/json"
